backend/requirements.txt
Flask==3.1.0
pymongo==4.11.3
flask-jwt-extended==4.7.1
bcrypt==4.2.1
python-dotenv==1.0.1
flask-cors==5.0.1
